[ADD] samosol: New field to print labels on invoice and sale order reports
Foreign customers require customizing the labels of their identification
types. Therefore, the client needs a field based on their identification
where they can modify the texts that will later be included in the
printed representations of the sales and billing documents.

With this change, Samosol can edit the new field partner_label with any
information and structure requested by its customer, and then this will
be printed in the PDF.

Issue reported on helpdesk.ticket #21471